Add ability to locally install `setup.py`.
This introduces a `setup.py` file for the command-line interface to
allow local installations of the package. This should allow to evaluate
the `isct` command from anywhere in the system. To allow the script to
be setup as executable, the main routine of `isct.py` has to receive no
required arguments.

This contributes towards #15 and #4.