Add events to YAML and support XML config files.
After initial YAML configuration of a patient and after evaluation of
the `virtual_patient_generation` the events are added to the
configuration file. Although no additional logic is present, it could
enable to create different pipelines per patient based on certain
properties in the patients configuration.

Additionally, the YAML configuration is converted towards XML to support
the current submodules that require the XML import formats. Which should
likely be changed (#12).