Refactor patient code; adds the `Patient` class.
The `Patient` class is introduced as an extension on the normal `dict`
instance. This attaches some often used functionality on top of the
dictionary, which enables easier interaction with patient configuration
files, as we can consistently deal with the `Patient` class rather then
to either deal with the directories of the patient or directories of the
patient configuration file.