Add a way to store git sha in config files.
Preliminary solution to store some version information into the
configuration files. This now stores a git sha of the installed package.
Note, this strongly relies on the fact that we install the package as
`--editable` using `pip`, which provides a link to the local repository.
Therefore, we can probe the git sha from that directory and store it
within the configuration files.

A more long term solution would be transition into a version-based
solution, where the package's version is directly linked to a specific
tag of the version controlled repository.

Closes #8.