Add `isct trial status`, `isct trial ls` commands.
Both commands utilise `tree` to report on the status of an trial. The
`trial status` commands illustrates the progress of all events of all
patients in the trial directory. For this, it is required to read each
patient's configuration file (which works when all is executed locally).

The `trial ls` commands performs similar function as `tree` for a
provided trial directory.