Provide directory filter to `tree`.
This replaces the `patient_only` argument with a generic function as
argument that receives a path. The boolean return value of this function
is then used to filter the files encountered in the `tree` traversal in
addition to the check if the path is an directory.

This allows to separate the utility from any patient's logic, as this is
now passed in externally. The `tree` routine is not concerned about
the directories being a patient yes or no, as it only evaluates the
provided filtering functions, which provides the required Patient logic
externally.