Adds `Container` class to enable Singularity.
To enable Singularity container to be used alongside the Docker
containers, the original code for the Docker containers has been
refactored. This introduces the `Container` base class that defines the
functionality required for container method to be used, which are
implemented separately for both Docker and Singularity (this one is not
complete yet).

The different logic per container type is now hidden in the `Container`
instance and the remaining workflow code does not need to worry about
dealing with either type of container, as it operates on these
`Container` instances. The commit adds functionality for Docker and
includes initial tests for the base `Container` class as well as the
Docker implementation.

Support for singularity is to be added.