Add `Vagrant` configuration file in /software/.
For `MACOS` the Singularity containers are build inside a `Vagrant` VM.
This requires the presence of a configuration file `Vagrantfile` to be
present in that location. The easiest solution is to place this file
inside /software/, as its directory is shared by default with the
`Vagrant` VM. Thus inside the VM, all submodule directories are
accessible.