Adds the image path to `Singularity`.
Compared to Docker, Singularity uses separate files of the containers.
These lives at any possible location on the host's system. Therefore, we
add a `path` argument to the `Singularity` class to keep track of this
path. Note, it is required for almost all operations to provide the
full, explicit path to the desired singularity image to run, which is
why this path is stored in the class itself. Also, as we cannot perform
any operation without this path, the `Singularity` class asserts this is
a valid and existing directory on initialisation.