Update command-line arguments for `Singularity`.
The `isct` commands now include a flag to use Singularity containers
rather than Docker containers. This requires the user to pass in a
directory where all the images are stored. Ultimately, these images can
be obtained either locally or downloaded from a remote location and
simply have to be assigned their corresponding directory on disk.