Schema validate optional `--singularity` argument.
The argument `--singularity=PATH` provides either a `PATH` or `None`
after parsing with `docopt`. Therefore, we cannot use an optional
argument from `schema` as the argument is always present. This adds
support to either allow `None` or to (attempt to) convert the path to a
string and check it exists.