Add documentation and enable `parallel` with remote server.
To enable `parallel` with remote server it is easier to write
relative paths for the patient commands. This ensures we only need
to mimic the relative directory structure on the remote system to
run the simulations.

Also `isct container build` now accepts directories that might
contain single files. Its schema only validates on existing paths
and will filter out anything that is not a directory. This makes
the interface somewhat more user-friendly as this filter is not
required when passing a directory structure to the command.