Add validation for patient configuration.
This adds functionality to validate the contents of a patient's YAML
configuration file. The contents are processed using a predefined schema
of the expected parameters. This allows to verify a number of properties
of the YAML file, e.g. the type of variables, the range of variables,
feasible values for restricted parameters.

With this functionality added to the `Patient` class, two modifications
are made to the set of `isct` commands. First, a `isct patient validate`
command is added. This evaluates and (attempts to) validate the
configuration files of the provided patient directories. This simply
reports to the console whether the encountered config files are valid or
not. Additionally, the `isct trial run` command has an additional flag
`validate` that will force a validation to be performed on the YAML
before running the simulation.