Update available parameters for UQ.
Rearranges the encoding of the parameters to be split per input files,
thus the parameters provided in `params` are added to the respective
input files, e.g. configuration files or default setting files in any of
the submodules. Remaining parameters that have not been inserted
anywhere, should trigger an error in the encoder.