Apply `YAMLDecoder` to extract all inputs.
Update to a more recent version of `EasyVVUQ` to import the
`YAMLDecoder` more easily and use this decoder to extract all the
parameters that were injected in the patient configuration file. This
should allow to extract all the input parameters directly from the
configuration files using a single decoder. The output variables of
interest are still to be extracted.