Extend notebook: analyses and database reloading.
This adds the QMC method, which is only a minor variation compared to
the implementation of PCE. Additionally, an example on storing/reloading
a database is given at the end of the notebook. This should allow to
efficiently share VVUQ results by only distributing the serialised
results.

Also adds some notes/TODOs for future extensions.