Update submodules and update eventmodule API.
This moves the submodules to newer commits including the updated
eventmodule. Additionally, this fixes the API calls to the modules, i.e.
the arguments for the eventmodule are updated from `handle_event` to
`event`. Finally, the default value for the clot length is updated to be
between the bounds as indicated by the thrombectomy module.