Update `Event, State` to track event states.
Some events are evaluated multiple times within a single pipeline
definition, e.g. the blood flow, perfusion, and cell death models are
evaluated for a `baseline`, `stroke` and `treatment` state. We want to
store the data of each of these states separately, both to compare the
simulation results between states as well as to prevent overwriting of
previously generated data.

The enumeration `State` is added that tracks these three states. All of
the possible states, i.e. `baseline, stroke, treatment`, are stored
within the patient configuration as a list of strings. These strings
have two functions: provide a list of all possible states to any module
and the individual strings represent the subdirectories where to store
the output. Thus, when in state `baseline` the output of the submodules
is to be stored in `/patient/baseline`, separate from the outputs in
`/patient/stroke` and `/patient/treatment`.

Each event contains an integer value indicating the current `State`.
This index points into the list of states given in the patient
configuration file. Currently, index `0` corresponds to state
`baseline`. This enables the different `API.py` implementations to
access previous states, simply by incrementing/decrementing the current
state value and by accessing the corresponding entry in of all states.