Patch file permissions with `--root`.
When running with `--root` the user indicates to have permissions to use
docker, e.g. by a user group. However, the files generated by docker
still have permissions of `root`. Without `--root` this is resolved by a
`sudo chown` operation, which is not possible in combination with the
`--root` flag.

To overcome this, we can query the evaluated docker container again to
perform the `chown` operation, attempting to change the permissions of
all files present in the bind mount volume `/patient` to be aligned with
the permissions of the host directory. As this directory was created
with the corresponding user's permissions, the `root` permissions set by
the container are overwritten to match the user's permissions again.

This resolves #59.