Set default log level to `warning` only.
Logging to the console is suppressed when no verbose flag `-v` is given
unless the logged message has a level of either warning, error, or
critical. By passing `-v` the console logging is updated to report
anything from the information level, i.e. the basic print statements.
Debug messages are still hidden and remain available in the logs written
to `/tmp/isct.log`.