Add `terminated` property to patient.
The `terminated` property of the patient indicates if the patient is
terminated and its containers should not be evaluated anymore. The
`terminated` property is set when a non-zero exit code is detected after
running any of the events, i.e. the container commands have non-zero
exit codes.

After such event, the `patient.terminate()` routine is called to
implicitly set `patient.terminated` to True, by inserting the
`terminated` key into the patient's directory. Although this typically
will have a value of `True` it can still be overwritten.

Once `patient.terminated` returns true, subsequent calls to any
containers will not be executed.