Remove files in copied directory by `ISCTEncoder`.
The `ISCTEncoder` copies a template directory to provide an already
initialised starting point for the VVUQ analysis. A problem occurs when
signalling files, e.g. `clot_present`, that are normally written during
the simulations, are already present from the start. This might cause
the models to choose the wrong branch at the start of the VVUQ.

To resolve this, a subset of the files are removed from the copied
directory. This should ensure that no signalling/result files are
present that influence the branching of the simulations, while still
allowing the models to reuse preprocessed results from the template
directory.