Redefine random seed at trial level.
The random seed used for the patient is now defined at the trial level
and stored inside the `trial.yml` file, where a copy is still stored
inside the specific `patient.yml` for reference. When creating a cohort
of patients it is this random seed, i.e. the one defined at the trial
level, that is passed towards the virtual-patient model. Thus, all
patients utilise the same random seed, where the randomness is
introduced by subsequent sampling of the model, rather than having an
unique, individual seed per patient.