Update vessel mapping towards blood-flow.
This updates the mapping from the integer values from the virtual
patient model towards a *fixed* vessel in the one-dimensional blood-flow
model. This ensures that the clot can find a matching vessel, although
the correct vessel is to be assigned.

Also, the left/right side of the clot is now determined as a random
choice between left/right with equal chance. Could be improved by a
specific distribution.