Replace underscores to hyphens when matching.
The naming convention of the submodules differs, where some use
underscores and others use hyphens. The code maps all underscores to
hyphens for the container names, thus the matching will also be
performed on strings where all underscores are mapped. The user can then
provide either the tag of the container, or the tag matching directory.