Either provide (`--criteria`) or (`-n`, `--seed`).
Trial create accepts either a criteria file by `--criteria` or the
optional `-n` or `--seed` arguments. This is to prevent overwriting
trial specific settings by the number of patients or the seed arguments,
as ideally these values are set directly in the given configuration file
containing the in/exclusion criteria.