Update log messages and passing criteria files.
If a criteria file is provided, we pass this file into the virtual
patient model. Otherwise, only the number of patients (i.e. sample size)
and the random seed are forwarded to the virtual patient model.

Note, when appending new patients to existing trials, first a temporary
configuration file is passed towards the virtual patient model. Only
after this command was completed successfully, the actual trial's
configuration is updated with the new sample size. This is to prevent
increasing the sample size, but the command failing, i.e. generating no
new patients.