Split VVUQ support into extra feature `isct[vvuq]`.
Allows to install `isct` without the dependencies for VVUQ analysis,
mainly without the requirements imposed by `EasyVVUQ`. To perform such
analysis, the package can be installed as `pip install isct[vvuq]` to
include the additional dependencies during installation.

To simplify testing for both versions of the package, support for `tox`
is also defined. By invoking `tox`, the tests are evaluated using the
`dev` dependencies with and without `vvuq`. The linter `flake8` is still
evaluated for all parts of the package.