Makefile for setup, installation, and testing.
A Makefile is added to setup the project, including the virtual
environment and required dependencies. `make install` will install
all dependencies, i.e. both for the `dev` and `vvuq` versions, where
a virtualenv is setup and activated before installing the packages.
Note, the `extras_require` are combined with the `[all]` flag,
indicating that all dependencies need to be installed in the
environment.

Additionally, `make test` invokes the defined tests through invoking
`tox` and steps through each defined environment. A minimal coverage
report is provided for each environment, where a html version is
produced for the complete environment including `vvuq` dependencies.

Finally, `make distclean` cleans the created files to resemble the
original state of the distribution before invoking `make install` or
`make test`.