Adopt click and refactoring `isct`.
This explores and works towards adopting `Click` to define the
command-line interface, while simultaneously refactors the backing
implementation of the `isct` functionality.

The main motivation to explore `Click` is the better support for nested
command groups and the simplified assertions on the passed arguments,
e.g. better enforcements of paths (exist, writable, ...) directly from
their definitions. This removes the need to perform any validation
checks on these argument, as previously done with `schema`.

Moreover, the `docopt`-style command-line interface was getting somewhat
clumsy with the nested groups, due to repeated definitions in each of
the docstrings. Any overlap with `Click` seems less abrasive, especially
as the corresponding documentation is generated automatically and is not
mixed with the implementation of the arguments.

Finally, `Click` seems to make testing easier with some helper functions
that clean up the testing code quite a bit.

The change of these command-line frameworks allows for a possibility to
refactor the backing code as well. Thus, the arrangement of the objects
that define the patients, trials, containers, etc., are updated too in
attempt to further simplify the original implementation. The main
changes that are introduced, are that both `Patient` and `Trial` derive
their overlapping dictionary behaviour from a single parent `Config`
class. Another update, is the introduction of various "runners".
These classes provide different types of behaviour of evaluating the
container commands, e.g. should the command be run in the terminal, just
provided over `stdout` (for instance for `GNU Parallel`), or just
logged using the logger.