Prevent overwriting directories and update `read`.
The `isct trial create` will not overwrite existing directories and does
not provide an option to do so anymore. It leaves removing and/or
renaming of these directories up to the user.

The class methods `read` are updated for `Patient` and `Trial` to allow
to more easily pass in the `runner` instances, i.e. directly when
initialising these objects rather then assigning afterwards.