Store container directory in `trial.yml`.
The container directory is stored inside the trial configuration when
the trial is created. Subsequent calls to the trial can pickup the
location of the containers from there. A `None` (in YAML `null`) value
indicates no container directory was given, i.e. Docker is used, whereas
a valid path indicates Singularity. If the path is non-existing, the
command-line will raise a usage error indicating the wrong container
path and suggesting the user to update it accordingly.