Make `Trial` iterable and display progress bar.
This implements the trial's `__iter__` routine, which yields a patient
instance for each patient considered in the specific trial. This enables
for direct iteration of the patients in a trial and, additionally,
allows to easily visualise a simple progress bar in the terminal when
running the virtual patients in the trial.

Note, the progress bar is is only displayed when the patients are
evaluated sequentially without raised verbosity. If the verbosity is
raised to `DEBUG` the log printing will conflict with the progress bar
and is therefore not displayed when logs are printed in the console.