Add the `isct trial list-key` command.
This command lists the values corresponding to the given key in all
patient configurations of the trial. Of the encountered values only the
unique entries are stored, where their occurrences are counted. The
command lists all values, and their count, by default. However,
providing the `-n` option flag allows to limit the listed values to only
the `n` most common values.