Return `container-path` as string.
This allows to set the container paths directly when updating any of the
configuration files, while prevents inserting `pathlib.Path` instances
into the dictionary. The problem with the latter is, that this does not
nicely expand into the YAML document, i.e. it is written as an object
that on reading is not defined. So, by ensuring `def container_path`
returns a string of the path, we avoid inserting objects.