Mark patient complete and `--skip-completed` flag.
The patients are marked as complete after all event simulations are
completed. For now this only aims to mark patients as completed when
running the trial using per-patient parallelism or sequential
evaluation, as this label is set to true after all events are passed
(sequentially).

The `isct trial run` command accepts a `--skip-completed` flag that
enables running only the simulation pipeline for virtual patients that
are marked as completed yet.