Add command-line flag to keep/clean large files.
This add functionality to the command-line interface to clean large
files during the simulation of large cohorts of virtual patients. The
flag `--keep-files/--clean-files` toggles this behaviour, where the
default is to keep the files. If the files are to be cleaned, all files
in any subdirectory of the patient larger than the
`isct.utilities.MAX_FILE_SIZE` threshold will be deleted. This ensures
(relatively) small disk usage of the resulting trial and nested patient
directories.

To implement this behaviour a `LowStoragePatient` is added as subclass
of the default `Patient` class.