Assert when a simulation event failed.
For evaluated commands using the local runner, it is possible to
determine if the command was evaluated successfully or not. If so, a
boolean is returned. In this scenario, a True value is asserted, such
that the pipeline is terminated early on a broken simulation event.

For runners that do not have such a notion of success, for instance when
simply logging to the screen/stdout, the return value of None is also
accepted in the assertion.