Update VVUQ and support missing trial.yml.
Updates the VVUQ notebook to a more recent version of `easyvvuq` as well
as adds support to running the trials when trial configuration files
might be missing. This is supported by falling back on attempting to run
simulations for all subdirectories of the trial, as long as those can
successfully be interpreted as patient directories. Additionally, the
patient runner is less strict about the filename conventions and will
adopt different directory names if needed (as is the case within VVUQ).