Add QCG-PilotJob support and desist command.
This adds support to run trials using `QCG-PilotJob` as a runner through
providing the `--qcg` flag to the `trial run` command. Using this
command invokes the `QCQRunner` to distribute and run the tasks. In this
way, the QCG manager is leveraged to submit the trials that need to be
run which are then, automatically, scheduled and distributed on the
available hardware. Once all jobs are completed, desist will return.
Note currently no retries are attempted for failed jobs.

This also enables for simpler running of trials across multiple systems,
e.g. when running on HPC environments. The distribution of the scheduled
tasks is handled by QCG only and we do not need to ensure that GNU
Parallel uses the right settings and environments across the different
nodes.