Avoid using LowStoragePatient on dry runs.
This avoids converting the Patient to a LowStoragePatient when the
dry-run flag is provided, such that the cleaning of the files is not
invoked when the patient is evaluated in a dry run. This now results in
only the commands to executed to be logged and the files to remain
present.