Add Runner.wait to simplify QCGRunner.
By including `wait` in the abstract base class `Runner` all runners will
have a `wait` function, although it lacks any implementation. This
simplifies the implementation of the `QCGRunner` as we do not need to
specialise manually on the type of runner to determine if the `wait`
attribute will be present.