Accept bind paths in `trial outcome --compare`.
This enables to pass paths as `source:destination` to provide the
complete bind path and override the internal default for the bind mount
of the comparison path. A simple check is performed to see if the path
can be interpreted as a bind path: it does not exist on the file system
and only a single colon (:) is present.